+++ src/de/bokeh/skred/red/PrimPutChar.java
@@ -15,7 +15,7 @@ public class PrimPutChar extends Function {
         c.eval();
         Node a2 = c.getTos();
         OutputPort p = (OutputPort) c.get1();
-        int ch = a2.intValue();
+        int ch = a2.intValue().intValue();
         p.getPrintStream().print((char)ch);
         Node result = Data.valueOf(0, Data.valueOf(0), Int.valueOf(0));
         c.pop3();
